refactor: Break up home component
