+++ client/src/components/Home/Home.styled.js
@@ -1,98 +1,10 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import {
-	$greyLightest,
-	$greyLighter,
-	$greyLight,
-	$greyDark,
-	$greyDarker,
-	$mintLight,
-	$elfGreen,
-	$mintDark,
-	$ash,
-} from '../../assets/styledVars';
+import { $greyDark, $mintLight, $elfGreen, $mintDark, $ash } from '../../assets/styledVars';
 
 export const Container = styled.div`
 	text-align: center;
 	overflow-x: hidden;
-`;
-
-/* =============================================
-									First Page
-============================================= */
-
-export const FirstPage = styled.div`
-	z-index: 0;
-	* {
-		z-index: 20;
-	}
-	position: relative;
-	width: 100%;
-	min-height: 100vh;
-	background: linear-gradient(15deg, ${$greyLighter}, #fff 80%);
-	padding: 160px 2rem 4rem;
-
-	.icon__logo {
-		height: 322px;
-		padding: 3rem 0;
-	}
-`;
-
-export const HomeNav = styled.div`
-	z-index: 20;
-	position: absolute;
-	top: 0;
-	left: 0;
-	height: 100px;
-	width: 100%;
-	display: flex;
-	justify-content: center;
-	padding: 0 2rem;
-
-	@media (max-width: 680px) {
-		font-size: 12px;
-	}
-`;
-
-export const NavContent = styled.div`
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	width: 100%;
-	max-width: 1024px;
-`;
-
-export const Logo = styled.div`
-	font-size: 24;
-	color: ${$greyDarker};
-	font-size: 1.5rem;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-
-	.icon__logo {
-		height: 60px;
-		padding-right: 1rem;
-	}
-`;
-
-export const NavLinks = styled.div``;
-
-export const NavLink = styled(Link)`
-	transition: 0.3s;
-	border-radius: 6px;
-	display: inline-block;
-	padding: 0.5rem 2rem;
-	text-decoration: none;
-	background: transparent;
-	${({ focused }) =>
-		focused && `border: 2px solid ${$mintLight}; &:hover{ background: ${$mintLight} }`};
-	color: ${$greyDark};
-	font-size: 1.125rem;
-	margin-left: 2rem;
-	&:hover {
-		color: ${$mintDark};
-	}
 `;
 
 export const LargeTitle = styled.h1`
@@ -132,56 +44,6 @@ export const SmallLink = styled.a`
 	font-size: 0.8rem;
 `;
 
-/* =============================================
-									Middle Section
-============================================= */
-
-export const MiddleSection = styled.div`
-	z-index: 0;
-	& * {
-		z-index: 20;
-	}
-	position: relative;
-	width: 100%;
-	min-height: 100vh;
-	background: ${$greyLightest};
-	padding: 5rem 2rem 0;
-
-	${LargeTitle} {
-		max-width: 600px;
-		margin: 0 auto;
-	}
-
-	${LargeSubtitle} {
-		margin: 3rem 0 0 0;
-	}
-
-	${SmallLink} {
-		display: inline-block;
-		margin: 1rem 0;
-		color: ${$greyLight};
-	}
-
-	#circle__one {
-		z-index: -10;
-		position: absolute;
-		left: 88%;
-		top: -100px;
-	}
-	#circle__two {
-		z-index: -9;
-		position: absolute;
-		left: 94%;
-		top: 135px;
-	}
-	#circle__three {
-		z-index: -10;
-		position: absolute;
-		right: 94%;
-		top: 335px;
-	}
-`;
-
 export const FlexRow = styled.div`
 	display: flex;
 	justify-content: space-between;
@@ -192,160 +54,5 @@ export const FlexRow = styled.div`
 
 	@media (max-width: 700px) {
 		display: block;
-	}
-`;
-
-export const HabitStages = styled(FlexRow)`
-	padding: 5rem 0 5.5rem;
-
-	@media (max-width: 700px) {
-		padding: 6.5rem 0 5.5rem;
-	}
-`;
-
-export const HabitStage = styled.div`
-	width: 170px;
-	h3 {
-		color: ${$ash};
-		font-weight: 400;
-		text-align: center;
-		padding: 1.5rem 0;
-	}
-	p {
-		font-size: 0.9rem;
-		color: ${$greyDark};
-		line-height: 1.75;
-	}
-	svg {
-		height: 99px;
-	}
-
-	@media (max-width: 700px) {
-		margin: 0 auto 3rem;
-		width: 50%;
-	}
-`;
-
-/* =============================================
-									Last Page
-============================================= */
-
-export const LastPage = styled.div`
-	z-index: 0;
-	& * {
-		z-index: 20;
-	}
-	position: relative;
-	min-height: 200vh;
-	overflow: hidden;
-	background: linear-gradient(${$greyLighter}, #fff 43%);
-	padding: 0 2rem;
-
-	${LargeTitle} {
-		margin: 6rem auto 1rem;
-	}
-	${LargeSubtitle} {
-		display: inline-block;
-		font-size: 2.125rem;
-		color: ${$ash};
-		max-width: 800px;
-		margin: 16rem auto 10rem;
-		z-index: 1;
-		position: relative;
-		svg {
-			position: absolute;
-			top: -188px;
-			left: -110px;
-			z-index: -1;
-		}
-	}
-	${RegisterButton} {
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		font-size: 1.5rem;
-		padding: 1rem;
-		height: 80px;
-		min-width: 500px;
-
-		@media (max-width: 650px) {
-			min-width: 90%;
-			width: 80%;
-		}
-	}
-
-	#circle__four {
-		z-index: -10;
-		position: absolute;
-		right: 94%;
-		top: 335px;
-	}
-	#circle__five {
-		z-index: -10;
-		position: absolute;
-		right: 88%;
-		bottom: 50px;
-	}
-	#circle__six {
-		z-index: -10;
-		position: absolute;
-		right: 83%;
-		bottom: -40px;
-	}
-	#circle__seven {
-		z-index: -10;
-		position: absolute;
-		left: 88%;
-		bottom: 120px;
-	}
-`;
-
-export const AppFeatures = styled(FlexRow)`
-	max-width: 770px;
-`;
-
-export const AppFeature = styled.div`
-	position: relative;
-	width: 210px;
-	text-align: left;
-	svg {
-		position: absolute;
-		top: -4px;
-		left: -21px;
-		width: 36px;
-	}
-	h3 {
-		display: inline-block;
-		color: ${$ash};
-		font-size: 1.5rem;
-		font-weight: 400;
-		padding-left: 2rem;
-	}
-	p {
-		margin-top: 2rem;
-		color: ${$greyDark};
-		font-size: 1.125rem;
-		line-height: 1.5;
-	}
-
-	@media (max-width: 700px) {
-		margin: 0 auto 4rem;
-		width: 66%;
-	}
-`;
-
-export const StyledScreenshot = styled.img`
-	display: block;
-	margin: 0 auto;
-	width: 100%;
-	max-width: 873px;
-`;
-
-export const FooterSmall = styled.p`
-	font-size: 0.75rem;
-	color: ${$greyDark};
-	padding: 3rem 0;
-	a {
-		color: ${$greyDark};
 	}
 `;
